+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # 0.3.0 - UNRELEASED
 
 - updated pykit@0.8.0
+- listed public exports
+- renamed throw_err_evt() -> throw()
+- added ServerBusCfg.are_errs_catchlogged which allows to catch+reraise thrown to the bus errs
+- removed ServerBusCfg.is_invoked_action_unhandled_errs_logged in favor of ServerBusCfg.are_errs_catchlogged
 
 # 0.2.3